pyproject.toml
[tool.poetry]
name = "jobman"
version = "0.4.0"
description = "A command line job manager with flexible support for retries, timeouts, logging, notifications, and more"
license = "MIT"
authors = ["Ryan Wallace <ryancswallace@gmail.com>"]
maintainers = ["Ryan Wallace <ryancswallace@gmail.com>"]
readme = "README.md"
homepage = "https://ryancswallace.github.io/jobman"
repository = "https://github.com/ryancswallace/jobman"
documentation = "https://ryancswallace.github.io/jobman"
keywords = ["python", "cli", "command-line", "background-jobs", "supervisor", "retry", "cli-app", "command-line-tool", "nohup"]
classifiers = [
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Topic :: Utilities"
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/ryancswallace/jobman/issues"

[tool.poetry.scripts]
jobman = 'jobman.cli:cli'

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
click = "^8.1.7"
psutil = "^5.9.5"
rich = "^13.5.2"
ruamel-yaml = "^0.17.32"
pydantic = "^2.3.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
autoflake = "^2.2.1"
black = "^23.7.0"
pyinstaller = "^5.13.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.1"
mypy = "^1.5.1"
pytest-cov = "^4.1.0"
types-psutil = "^5.9.5.16"
tox = "^4.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=jobman"
testpaths = [
    "tests",
]

[tool.mypy]
mypy_path = "jobman"
no_implicit_optional = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_defs  = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
allow_redefinition = false
local_partial_types = true
strict = true
disallow_incomplete_defs = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
warn_unused_configs = true
disallow_any_unimported = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_any_decorated = false
disallow_any_explicit = false
disallow_any_expr = false
